paging
mostly testing

+++ riscv64/src/memory.rs
@@ -2,6 +2,6 @@ use crate::platform::PHYSICAL_MEMORY_OFFSET;
 
 /// Convert physical address to virtual address
 #[inline]
-pub const fn phys_to_virt(paddr: u64) -> u64 {
+pub const fn phys_to_virt(paddr: usize) -> usize {
     PHYSICAL_MEMORY_OFFSET + paddr
 }
